+++ src/main/java/io/nagurea/smsupsdk/lists/create/CreateListService.java
@@ -1,7 +1,7 @@
 package io.nagurea.smsupsdk.lists.create;
 
 import com.google.gson.Gson;
-import io.nagurea.smsupsdk.common.post.POSTSMSUpService;
+import io.nagurea.smsupsdk.common.http.post.POSTSMSUpService;
 import io.nagurea.smsupsdk.helper.json.GsonHelper;
 import io.nagurea.smsupsdk.lists.create.arguments.CreateListArguments;
 import io.nagurea.smsupsdk.lists.create.body.CreateListBody;